+++ docs/release-notes.md
@@ -5,8 +5,9 @@
 ## Upcoming release
 
 ### For players:
-* Improved mod load times.
 * The pre-installed `Essentials` mod is now renamed to `Nickel.Essentials`. **Be sure to remove the old `Essentials` folder when updating Nickel.**
+* Nickel now lets you resize its window.
+* Improved mod load times.
 * Added `Nickel.ModSettings`, creating a common interface for all mods to add their settings to.
 * Added various settings for all Nickel modules.
 * Update checks now remind you to set them up properly, and when your API key / token gets revoked or expires.